requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-heroku==0.3.1
django-modeltranslation==0.17.3
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
drf-yasg==1.20.0
google-auth==2.3.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pdfkit==1.0.0
phonenumbers==8.12.42
Pillow==8.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==3.0.6
python-pdf==0.39
pytz==2021.3
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
wkhtmltopdf-pack==0.12.3.0
